// src/shared/api/models/getApiInsightsEmployeePaymentsType200.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * API
 * OpenAPI spec version: 1.0.0
 */
import type { GetApiInsightsEmployeePaymentsType200Data } from './getApiInsightsEmployeePaymentsType200Data';

export type GetApiInsightsEmployeePaymentsType200 = {
    data: GetApiInsightsEmployeePaymentsType200Data;
    message: string;
};