// src/shared/api/models/getApiInsightsEmployeePaymentsType200DataEmployeesItem.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * API
 * OpenAPI spec version: 1.0.0
 */
import type { GetApiInsightsEmployeePaymentsType200DataEmployeesItemPaymentMethodsItem } from './getApiInsightsEmployeePaymentsType200DataEmployeesItemPaymentMethodsItem';
import type { GetApiInsightsEmployeePaymentsType200DataEmployeesItemSummary } from './getApiInsightsEmployeePaymentsType200DataEmployeesItemSummary';

export type GetApiInsightsEmployeePaymentsType200DataEmployeesItem = {
    employeeId: number;
    employeeName: string;
    paymentMethods: GetApiInsightsEmployeePaymentsType200DataEmployeesItemPaymentMethodsItem[];
    summary: GetApiInsightsEmployeePaymentsType200DataEmployeesItemSummary;
};